requirements_.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
azure-core==1.25.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==38.0.1
datasets==2.4.0
deepface==0.0.75
defusedxml==0.7.1
dill==0.3.5.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-cleanup==6.0.0
django-filter==22.1
djangorestframework==3.13.1
filelock==3.7.1
fire==0.4.0
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.1
fsspec==2022.7.1
gast==0.4.0
gdown==4.5.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
msrest==0.7.1
mtcnn==0.1.1
multidict==6.0.2
multiprocess==0.70.13
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
retina-face==0.0.12
rsa==4.9
sentencepiece==0.1.97
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
